Rename tradfri base entity module
