+++ homeassistant/components/tradfri/entity.py
